Move useFrontmatter one level up
